updated ChatScreen and CustomActions
added an array of objects to onSend in CustomActions, now they should be complete. Deleted all setImage/setLocation instances, deleted the View around MapView. Result: location can be loaded on ios. On android, finally the library can be accessed to pick an existing image and briefly a permission window shows up but disappears.